ARM: Do not enable ARM_DMA_MEM_BUFFERABLE for some RealView boards
RealView boards with certain revisions of the L210/L220 cache controller
may have issues with recent changes to the mb() barrier implementation
(DSB followed by an L2 cache sync).

This patch proposes that the mb() barrier performs an outer cache sync
only when ARM_DMA_MEM_BUFFERABLE is enabled. When this is option is
disabled, the coherent DMA buffers are mapped as Strongly Ordered and
there is no need for an L2 cache sync.

The patch also disables ARM_DMA_MEM_BUFFERABLE for some of the RealView
boards with L210/L220.

Signed-off-by: Catalin Marinas <catalin.marinas@arm.com>
Cc: Russell King <rmk@arm.linux.org.uk>
Cc: Linus Walleij <linus.ml.walleij@gmail.com>
Cc: Bjoern B. Brandenburg <bbb.lst@gmail.com>