UBUNTU: reduce disk usage during buildd builds
BugLink: http://bugs.launchpad.net/bugs/645653

For each flavour build we keep all of intermediate trees intact.
This allows simple incremental updates in a development environment but
leads to very large disk footprints on the build.  For Maverick with three
flavours we are seeing 19GB for i386 and 25GB for amd64.  Anything over
15GB is likely to run the buildds out of disk and fail the build.

When building on a buildd (in full_build=true mode) clean out
the main build tree and the package tree for the debugging deb.
This reduces consumption to more like 10GB maximum regardless of
the flavour count.

Signed-off-by: Andy Whitcroft <apw@canonical.com>