UBUNTU: [Config] Set physical start and alignment 1M for virtual i386
BugLink: http://bugs.launchpad.net/bugs/710754

This also sets the PHYSICAL_ALIGNMENT to 16M for normal i386 flavours
because this is the default since around 2.6.31 and we just happened to
carry an older default value. But a recent update of the configuration
changed the PHYSICAL_START to the new default of 16M. So now both values
are in sync again.

That said, unfortunately there seems to be a problem with pv-grub on
i386 when the physical start is set to 16M which prevents booting the
t1.micro instances on EC2 (again). Therefor the physical start and
alignment is changed back for the virtual flavour of i386.

Signed-off-by: Stefan Bader <stefan.bader@canonical.com>
Signed-off-by: Tim Gardner <tim.gardner@canonical.com>