UBUNTU: [Config] Build some framebuffer drivers as modules for powerpc kernels.
At some point in the past with config update/consolidation, some framebuffer
drivers got built into the powerpc kernels. This config change builds these
drivers as modules, as the OpenFirmware framebuffer driver is already in the
kernel. This also allows DRM modules to take priority in the case of radeon
cards.

Signed-off-by: Luke Yelavich <themuso@ubuntu.com>
Signed-off-by: Tim Gardner <tim.gardner@canonical.com>
Signed-off-by: Leann Ogasawara <leann.ogasawara@canonical.com>