SAUCE: ARM: OMAP: Overo: remove duplicate call to overo_ads7846_init()
overo_ads7846_init() is already called from overo_spi_init(), and
calling it twice is not only unnecessary but causes a warning as
"reg-fixed-voltage.1" is already added to the device tree.

This has the nice side effect of fixing an overo reboot bug.
Buglink: https://bugs.launchpad.net/linux-linaro/+bug/830785

Signed-off-by: Daniel Mack <zonque at gmail.com>