UBUNTU: SAUCE: HID: magicmouse: Set resolution of touch surfaces
Add touch surface resolution information. The size of the touch surfaces
has been determined to the hundredth of a mm.

This patch has been accepted upstream and should land in 3.1.0. Touch
surface resolution is becoming important for our stack. This patch does
not change any existing functionality, but merely fills in the
resolution details.

Cc: Jiri Kosina <jkosina@suse.cz>
Cc: Michael Poole <mdpoole@troilus.org>
Cc: linux-input@vger.kernel.org
Cc: linux-kernel@vger.kernel.org
Signed-off-by: Chase Douglas <chase.douglas@canonical.com>
[jkosina@suse.cz: update comments and commit message]
Signed-off-by: Jiri Kosina <jkosina@suse.cz>
Signed-off-by: Leann Ogasawara <leann.ogasawara@canonical.com>