UBUNTU: [Config] standardise on HZ=250
For a couple of releases we have had an inconsistant HZ for amd64, this
has been reviewed a number of times and been deferred pending a more
thorough review.  At this stage it seem safer to simply standardise on
a common HZ=250 across all architectures which can support it until any
review is completed.

Signed-off-by: Andy Whitcroft <apw@canonical.com>
Acked-by: Tim Gardner <tim.gardner@canonical.com>
Signed-off-by: Leann Ogasawara <leann.ogasawara@canonical.com>