ENGR00170901-2 IMX/UART : add a wait queue for DMA RX in SMP
The DMA RX thread may run on another CPU, while the application
runs on one CPU.

The application will close the uart when it finishes the required reading.
But the DMA RX thread maybe still running in this situation.

So add a wait queue to synchronize the __TWO__ sides in the SMP.

Signed-off-by: Huang Shijie <b32955@freescale.com>
SIgned-off-by: Haitao Zhang <haitao.zhang@linaro.org>